requirements.txt
ffmpeg-python==0.2.0
Pillow==10.0.1
tk==0.1.0